Rename files and fix some conditions
