+++ back/src/maps/views.py
@@ -1,24 +1,21 @@
-import requests
-import xml.etree.ElementTree as xml
-import json
-
-from .helpers import build_url, build_xml
-from .validators import suggester, rev_geocoder, geocoder, check_addr
+from .external_api.maps_api import suggester, rev_geocoder, geocoder, check_addr
+from .external_api.osm_api import create_note, create_object
+from django.http import JsonResponse, HttpResponseBadRequest, HttpResponse
+from django.views.decorators.http import require_http_methods
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.csrf import csrf_exempt
+from django.shortcuts import get_object_or_404
+from django.core.serializers import serialize
 from .models import Object
 from django import forms
-from django.views.decorators.csrf import csrf_exempt
-from django.conf import settings
-from django.http import JsonResponse, HttpResponseBadRequest
-from django.views.decorators.http import require_http_methods
-from requests_oauthlib import OAuth1
-from django.core.serializers import serialize
+import json
 
 
 @csrf_exempt
 @require_http_methods("POST")
 def create_point(request):
     try:
-        data = (json.loads(request.body))
+        data = json.loads(request.body)
     except json.decoder.JSONDecodeError:
         return HttpResponseBadRequest('Wrong request count of fields - bad fields')
     if data.get('address') is None:
@@ -52,120 +49,10 @@ def create_point(request):
         note_osm_id = int(data['info']['id'])
         if request.session.session_key is not None:
             session = request.session
-            session['points'].append((point_db_id, note_osm_id))
+            session['points'].append(point_db_id)
             session.save()
     # TODO unite obj_osm_id with point_db_id
     return JsonResponse(data)
-
-
-def create_note(lat, lon, address_obj):
-    url = build_url(
-        getattr(settings, 'OSM_URL'),
-        getattr(settings, 'OSM_URL_NOTE'),
-        {
-            'lat': lat,
-            'lon': lon,
-            'text': address_obj['address'],
-        }
-    )
-    response = requests.post(url)
-    if not response.ok:
-        return {'status_osm': False}
-    data_json = {'status_osm': True}
-    try:
-        root = xml.fromstring(response.content)
-        data_json['info'] = {}
-        for child in root[0]:
-            if child.tag != 'comments':
-                data_json['info'][str(child.tag)] = child.text
-    except (IndexError, xml.ParseError):
-        return {'status_osm': False}
-    return data_json
-
-
-def create_object(lat, lon, address_obj, user):
-    extra_data = user.social_auth.get(provider='openstreetmap').extra_data
-    access_token = extra_data['access_token']
-    address = address_obj['address_details']
-    auth = OAuth1(
-        getattr(settings, 'SOCIAL_AUTH_OPENSTREETMAP_KEY'),
-        getattr(settings, 'SOCIAL_AUTH_OPENSTREETMAP_SECRET'),
-        access_token['oauth_token'],
-        access_token['oauth_token_secret']
-    )
-
-    url = build_url(
-        getattr(settings, 'OSM_URL'),
-        getattr(settings, 'OSM_URL_OBJECT_CREATE'),
-    )
-    data = build_xml({
-        'osm': {
-            'changeset': {
-                '_tags_': {'version': '0.6', 'generator': 'iD'},
-                'tag': {'_tags_': [
-                    {'k': 'comment', 'v': 'SqS'},  # TODO field comment should be generated by users according OSM API
-                    {'k': 'created_by', 'v': 'Open-FIAS v.1.0'},
-                    {'k': 'host', 'v': 'https://open-fias.ru'},
-                    {'k': 'locale', 'v': 'ru'},
-                    {'k': 'changesets_count', 'v': '1'}
-                ]}
-            }
-        }
-    })
-    response = requests.put(url, data=data, headers={"Content-Type": "text/xml"}, auth=auth)
-    if not response.ok:
-        return {'status_osm': False}
-    commit_id = int(response.content)
-
-    url = build_url(
-        getattr(settings, 'OSM_URL'),
-        getattr(settings, 'OSM_URL_OBJECT_UPLOAD').format(commit_id),
-    )
-    data = build_xml({
-        'osmChange': {
-            '_tags_': {'version': 0.6, 'generator': 'iD'},
-            'create': {
-                'node': {
-                    '_tags_': {'id': -1, 'lon': lon, 'lat': lat, 'version': 0, 'changeset': commit_id},
-                    'tag': {'_tags_': [
-                        {'k': 'name', 'v': address_obj.get('name')},
-                        {'k': 'addr:country', 'v': address.get('country')},
-                        {'k': 'addr:region', 'v': address.get('region')},
-                        {'k': 'addr:subregion', 'v': address.get('subregion')},
-                        {'k': 'addr:city', 'v': address.get('locality')},
-                        {'k': 'addr:suburb', 'v': address.get('suburb')},
-                        {'k': 'addr:street', 'v': address.get('street')},
-                        {'k': 'addr:housenumber', 'v': address.get('building')},
-                        {'k': 'addr:postcode', 'v': address_obj.get('postalcode')}
-                    ]}
-                }
-            },
-            'modify': {},
-            'delete': {
-                '_tags_': {'if-unused': True}
-            }
-        }
-    })
-    response_info = requests.post(url, data=data.encode('utf-8'), headers={"Content-Type": "text/xml"}, auth=auth)
-    if not response_info.ok:
-        return {'status_osm': False}
-
-    url = build_url(
-        getattr(settings, 'OSM_URL'),
-        getattr(settings, 'OSM_URL_OBJECT_CLOSE').format(commit_id),
-    )
-    response = requests.put(url, headers={"Content-Type": "text/xml"}, auth=auth)
-    if not response.ok:
-        return {'status_osm': False}
-    data_json = {'status_osm': True}
-    try:
-        root = xml.fromstring(response_info.content)
-        data_json['node'] = {}
-        for child in root[0].attrib.items():
-            data_json['node'][str(child[0])] = int(child[1])
-    except (IndexError, xml.ParseError):
-        return {'status_osm': False}
-    return data_json
 
 
 def send_db(lat, lon, address_obj, user):
@@ -173,8 +60,8 @@ def send_db(lat, lon, address_obj, user):
     form = ObjectForm(data=address_details)
     if form.is_valid():
         object_new = form.save(commit=False)
-        object_new.latitude = float(lat)
-        object_new.longitude = float(lon)
+        object_new.latitude = lat
+        object_new.longitude = lon
         if 'rank' in address_obj:
             object_new.rank = int(address_obj['rank'])
         if 'postalcode' in address_obj:
@@ -202,6 +89,49 @@ class ObjectForm(forms.ModelForm):
             'rank',
             'postcode',
         ]
+
+
+@csrf_exempt
+@login_required
+@require_http_methods("POST")
+def add_points_from_cookie(request):
+    if request.session.session_key is not None:
+        session = request.session
+        if 'points' not in session:
+            return HttpResponseBadRequest('Wrong request data - there are no points')
+        if len(session['points']) == 0:
+            return HttpResponseBadRequest('Wrong request data - there are no points')
+        try:
+            data = json.loads(request.body)
+            points = list(data['points'])
+        except (json.decoder.JSONDecodeError, KeyError, ValueError):
+            return HttpResponseBadRequest('Wrong request count of fields - bad fields')
+
+        for item in points:
+            if 'points' not in session:
+                return HttpResponseBadRequest('Wrong request data - should be session')
+            if item not in session['points']:
+                return HttpResponseBadRequest('Wrong request data - not correct id')
+            obj = get_object_or_404(Object, id=item)
+            obj.author = request.user
+            obj.save()
+        del session['points']
+        return HttpResponse(True)
+    return HttpResponseBadRequest('Wrong request data - should be session')
+
+
+@csrf_exempt
+def get_list_last_points(request):
+    points = Object.objects.all().filt_del(request.user).order_by("-created")[:10]
+    data = {'points': serialize('json', points)}
+    return JsonResponse(data)
+
+
+@csrf_exempt
+def get_list_points(request):
+    points = Object.objects.all().filt_del(request.user)
+    data = {'points': serialize('json', points)}
+    return JsonResponse(data)
 
 
 @csrf_exempt
